// plugins/testkube/src/types/openapi.ts
export interface components {
  schemas: {
// ...
    TestWorkflowExecutionSummary: {
      /**
       * Format: bson objectId
       * @description unique execution identifier
       * @example 62f395e004109209b50edfc1
       */
      id: string;
      /**
       * Format: bson objectId
       * @description identifier for group of correlated executions
       * @example 62f395e004109209b50edfc1
       */
      groupId?: string;
      /** @description identifier of the runner where it has been executed */
      runnerId?: string;
      /**
       * @description execution name
       * @example some-workflow-name-1
       */
      name: string;
      /** @description sequence number for the execution */
      number?: number;
      /**
       * Format: date-time
       * @description when the execution has been scheduled to run
       */
      scheduledAt?: string;
      /**
       * Format: date-time
       * @description when the execution result's status has changed last time (queued, passed, failed)
       */
      statusAt?: string;
      result?: components['schemas']['TestWorkflowResultSummary'];
      workflow: components['schemas']['TestWorkflowSummary'];
      tags?: components['schemas']['TestWorkflowTagValue'];
      /** @description running context for the test workflow execution (Pro edition only) */
      runningContext?: components['schemas']['TestWorkflowRunningContext'];
      configParams?: components['schemas']['TestWorkflowExecutionConfig'];
      /** @description generated reports from the steps, like junit */
      reports?: components['schemas']['TestWorkflowReport'][];
      resourceAggregations?: components['schemas']['TestWorkflowExecutionResourceAggregationsReport'];
    };
// ...
  };
// ...
}